Add in the rest of the data and fix card styling.
