--- src/assets/data/AdventurersTest.json
@@ -1,10 +1,0 @@
-[
-    {
-            "id": 1, 
-            "name": "Aisling Dryks", 
-            "locked": true , 
-            "baseStats": { "health": 12, "defense": 7, "movement": 6},
-            "skillStats": { "presence": 6, "lore": 3, "agility": 5, "perception": 2, "strength": 2 }
-    }
-]
-
