+++ src/components/character/CharacterGrid.tsx
@@ -1,7 +1,7 @@
 import { useEffect } from "react";
 import { Input } from "@/components/ui/input";
 import { ScrollArea } from "@/components/ui/scroll-area";
-import { useCharacterStore } from "@/stores/character_new";
+import { useCharacterStore } from "@/stores/character";
 import CharacterCard from "@/components/character/CharacterCard";
 import type { Character } from "@/types/character";
 
@@ -23,16 +23,18 @@ export default function CharacterGrid() {
     if (error) return <div className="text-red-500">{error}</div>
 
     return (
-        <div className="flex flex-col h-full gap-4">
-            <Input
-                placeholder="Search characters..."
-                value={filter}
-                onChange={(e) => setFilter(e.target.value)}
-                className="w-1/3"
-            />
-            <div className="flex-1 bg-muted rounded-lg shadow-inner p-4">
-                <ScrollArea className="h-full pr-4">
-                    <div className="grid grid-cols-1 md:grid-cols-2 lg:grid-cols-3 gap-4">
+        <div className="flex flex-col h-full">
+            <div className="sticky top-0 z-10 bg-background border-b p-4 flex gap-2 items-center">
+                <Input
+                    placeholder="Search characters..."
+                    value={filter}
+                    onChange={(e) => setFilter(e.target.value)}
+                    className="w-1/3"
+                />
+            </div>
+            <div className="flex-1 p-4">
+                <ScrollArea className="h-[calc(100vh-5rem)] pr-4">
+                    <div className="grid grid-cols-1 md:grid-cols-2 lg:grid-cols-3">
                         {filteredCharacters().map((char:Character) => (
                             <CharacterCard character={char} />
                         ))}
